vcodec: Add vcodec dvfs and pmqos
Add video codec dvfs and pmqos emi bw request

CR-Id: ALPS05459488
Feature: [Module]Video Driver

+++ drivers/media/platform/mtk-vcodec/vcodec_dvfs.h
@@ -38,8 +38,8 @@ struct codec_job {
 };
 
 struct codec_freq {
-	int freq[MODULE];
-	int active_freq;
+	unsigned long freq[MODULE];
+	unsigned long active_freq;
 };
